1 insert realizado com sucesso

+++ Application/Interface/IAssociadoAppService.cs
@@ -1,14 +1,12 @@
 ﻿using Application.ViewModel.Request;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Application.Interface
 {
     public interface IAssociadoAppService
     {
 
-        int PostCadastroAssociado(PostAssociadoViewModel postAssociadoView);
+        Int64 PostCadastroAssociado(PostAssociadoViewModel postAssociadoView);
 
         int GetCadastroAssociado(int id);
     }
